rename `thread_pool.rs` to `registry.rs`
